:recycle: refactor: 소개 페이지 각 component 및 style props type 명시
